+++ CHANGELOG.md
@@ -30,6 +30,8 @@ Copy and pasting the git commit messages is __NOT__ enough.
 - `EntryTactic` derived classes now contain `start_time`.
 - `info` returned by `hiway-v1` in `reset()` and `step()` methods are unified.
 - Changed instances of `hiway-v0` and `gym` to use `hiway-v1` and `gymnasium`, respectively.
+- `RoadMap.Route` now optionally stores the start and end lanes of the route.
+- `DistToDestination` metric now adds lane error penalty when agent terminates in different lane but same road as the goal position. 
 ### Deprecated
 - `visdom` is set to be removed from the SMARTS object parameters.
 - Deprecated `start_time` on missions.